refactor(core): move error-sanitizer to core/errors-sanitize
Move sanitizeErrorMessage and sanitizeErrorDetails from
src/features/liferay/errors/error-sanitizer.ts to
src/core/errors-sanitize.ts so env/worktree/deploy error
factories (PR16-19) can import shared sanitization logic
without coupling to liferay features.

src/features/liferay/errors/error-sanitizer.ts is kept as a
deprecated re-export facade for backward compatibility.
